1.6 pre12

+++ src/main/java/icyllis/modernui/ui/master/ModernScreen.java
@@ -65,27 +65,27 @@ public final class ModernScreen extends Screen {
 
     @Override
     public final void mouseMoved(double mouseX, double mouseY) {
-        manager.sMouseMoved(mouseX, mouseY);
+        manager.sMouseMoved((int) mouseX, (int) mouseY);
     }
 
     @Override
     public boolean mouseClicked(double mouseX, double mouseY, int mouseButton) {
-        return manager.sMouseClicked(mouseX, mouseY, mouseButton);
+        return manager.sMouseClicked((int) mouseX, (int) mouseY, mouseButton);
     }
 
     @Override
     public boolean mouseReleased(double mouseX, double mouseY, int mouseButton) {
-        return manager.sMouseReleased(mouseX, mouseY, mouseButton);
+        return manager.sMouseReleased((int) mouseX, (int) mouseY, mouseButton);
     }
 
     @Override
     public boolean mouseDragged(double mouseX, double mouseY, int mouseButton, double deltaX, double deltaY) {
-        return manager.sMouseDragged(mouseX, mouseY, deltaX, deltaY);
+        return manager.sMouseDragged((int) mouseX, (int) mouseY, deltaX, deltaY);
     }
 
     @Override
     public boolean mouseScrolled(double mouseX, double mouseY, double delta) {
-        return manager.sMouseScrolled(mouseX, mouseY, delta);
+        return manager.sMouseScrolled((int) mouseX, (int) mouseY, delta);
     }
 
     @Override
